+++ index.html
@@ -228,6 +228,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./src/app.js"></script>
+    <script type="module" src="./src/bootstrap.js"></script>
   </body>
 </html>
