+++ src/lazy-resources.js
@@ -1,0 +1,6 @@
+const onInitialLoad = 'requestIdleCallback' in window ? window.requestIdleCallback : cb => setTimeout(cb, 1000);
+
+onInitialLoad(() => {
+  import('./components/Dialog.js');
+  import('./components/CountryInfo.js');
+});
